chore(ma): timezone - refactor structure
