+++ micro-apps/@bundler/src/wc/index.wc.ts
@@ -2,7 +2,7 @@ import r2wc from "@r2wc/react-to-web-component"
 import i18nInit from "@windows98/i18n/configs/react.config"
 import { CalculatorView } from "../../../src/ma-calculator/src/ui/views"
 import { CalendarView } from "../../../src/ma-calendar/src/ui/views/calendar.view"
-import { TimezoneView } from "../../../src/ma-timezone/src/ui/views/timezone.view"
+import { TimezoneView } from "../../../src/ma-timezone/src/ui/views"
 import { TimerView } from "../../../src/ma-timer/src/ui/views"
 
 const MACalculator = r2wc(CalculatorView)
