+++ micro-apps/src/ma-timezone/src/data/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from "./timezone.repository"
