+++ micro-apps/src/ma-timezone/src/data/repositories/timezone.repository.ts
@@ -1,7 +1,7 @@
-import type TimezoneRepositoryContract from "../../domain/contracts/repository.contract"
-import type { TTimezoneValues } from "../../domain/models/timezone.model"
+import type { TimezoneRepositoryContract } from "../../domain/contracts"
+import type { TTimezoneValues } from "../../domain/models"
 
-export default class TimezoneRepository implements TimezoneRepositoryContract {
+export class TimezoneRepository implements TimezoneRepositoryContract {
 	public getTimezones(): Array<TTimezoneValues> {
 		return [-12, -11, -10, -9, -8, -7, -6, -5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
 	}
