+++ micro-apps/src/ma-timezone/src/domain/contracts/index.ts
@@ -1,0 +1,2 @@
+export * from "./domain.contract"
+export * from "./repository.contract"
