+++ micro-apps/src/ma-timezone/src/domain/domains/index.ts
@@ -1,0 +1,1 @@
+export * from "./timezone.domain"
