+++ micro-apps/src/ma-timezone/src/ui/modules/timezone-switcher/index.ts
@@ -1,0 +1,1 @@
+export * from "./timezone-switcher"
