+++ micro-apps/src/ma-timezone/src/ui/modules/wrapper/index.ts
@@ -1,0 +1,1 @@
+export * from "./wrapper"
