+++ micro-apps/src/ma-timezone/src/ui/modules/wrapper/use-wrapper.hook.ts
@@ -1,6 +1,6 @@
 import { useEffect, useMemo } from "react"
 import { useTranslation } from "react-i18next"
-import useTimezone from "../../hooks/use-timezone.hook"
+import { useTimezone } from "../../hooks"
 
 export function useWrapper() {
 	const { t } = useTranslation()
