+++ micro-apps/src/ma-timezone/src/ui/modules/wrapper/wrapper.tsx
@@ -1,7 +1,7 @@
-import { observer } from 'mobx-react-lite'
+import { observer } from "mobx-react-lite"
 import { DSFrame } from "@windows98/design-system"
 import { useWrapper } from "./use-wrapper.hook"
-import { TimezoneSwitcher } from '../timezone-switcher/timezone-switcher'
+import { TimezoneSwitcher } from "../timezone-switcher"
 
 export const Wrapper = observer(() => {
 	const { title } = useWrapper()
