+++ micro-apps/src/ma-timezone/src/ui/views/index.ts
@@ -1,0 +1,1 @@
+export * from "./timezone.view"
